+++ app/Core/Resource/Mail.php
@@ -4,14 +4,52 @@ namespace Solital\Core\Resource;
 
 class Mail
 {
-
+    /**
+     * Email sender
+     * @var string
+     */
     private static $sender;
-    private static $recipient;
-    private static $subject;
-    private static $message;
 
+    /**
+     * Email recipient
+     * @var string
+     */
+    private static $recipient;
+
+    /**
+     * Email subject
+     * @var string
+     */
+    private static $subject;
+
+    /**
+     * Email message
+     * @var string
+     */
+    private static $message;
+    
+    /**
+     * Send email
+     * @param string $sender
+     * @param string $recipient
+     * @param string $subject
+     * @param string $message
+     * @param string $reply_to
+     * @param string $type
+     * @param string $charset
+     * @param string $priority
+     */
     public static function send(string $sender, string $recipient, string $subject, string $message, string $reply_to = null, string $type = "text/plan", string $charset = "UTF-8", int $priority = 3)
     {
+        $validateSender = self::validateEmail($sender);
+        $validateRecipient = self::validateEmail($recipient);
+
+        if ($validateSender == false) {
+            return false;
+        } elseif ($validateRecipient == false) {
+            return false;
+        }
+
         self::$sender = $sender;
         self::$recipient = $recipient;
         self::$subject = $subject;
@@ -27,6 +65,21 @@ class Mail
         $send = mail(self::$recipient, self::$subject, self::$message, $headers);
         
         if($send) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    /**
+     * Validate the email
+     * @param string $email
+     */
+    public static function validateEmail(string $email)
+    {
+        $email = filter_var($email, FILTER_VALIDATE_EMAIL);
+        
+        if ($email) {
             return true;
         } else {
             return false;